requirements.txt
schedule
requests
pocketbase
python-dotenv
scikit-learn